atmosphere: Extract abstract base class for heterogeneous atmospheres
