+++ eradiate/scenes/atmosphere/_particles.py
@@ -4,7 +4,6 @@ Particle layers.
 from __future__ import annotations
 
 import pathlib
-import tempfile
 import typing as t
 import warnings
 
@@ -16,7 +15,7 @@ import xarray as xr
 
 import eradiate
 
-from ._core import Atmosphere, atmosphere_factory, write_binary_grid3d
+from ._core import AbstractHeterogeneousAtmosphere, atmosphere_factory
 from ._particle_dist import (
     ParticleDistribution,
     UniformParticleDistribution,
@@ -25,15 +24,12 @@ from ._particle_dist import (
 from ..core import KernelDict
 from ... import path_resolver
 from ..._mode import ModeFlags
-from ..._util import onedict_value
 from ...attrs import AUTO, documented, parse_docs
 from ...ckd import Bindex
 from ...contexts import KernelDictContext, SpectralContext
 from ...exceptions import OverriddenValueWarning, UnsupportedModeError
-from ...kernel.transform import map_cube, map_unit_cube
 from ...units import to_quantity
 from ...units import unit_context_config as ucc
-from ...units import unit_context_kernel as uck
 from ...units import unit_registry as ureg
 from ...validators import is_positive
 
@@ -41,7 +37,7 @@ from ...validators import is_positive
 @atmosphere_factory.register(type_id="particle_layer")
 @parse_docs
 @attr.s
-class ParticleLayer(Atmosphere):
+class ParticleLayer(AbstractHeterogeneousAtmosphere):
     """
     Particle layer scene element [``particle_layer``].
 
@@ -153,28 +149,6 @@ class ParticleLayer(Atmosphere):
         type="str",
     )
 
-    albedo_filename: str = documented(
-        attr.ib(
-            default="albedo.vol",
-            converter=str,
-            validator=attr.validators.instance_of(str),
-        ),
-        doc="Name of the albedo volume data file.",
-        type="str",
-        default='"albedo.vol"',
-    )
-
-    sigma_t_filename: t.Optional[str] = documented(
-        attr.ib(
-            default="sigma_t.vol",
-            converter=str,
-            validator=attr.validators.instance_of(str),
-        ),
-        doc="Name of the extinction coefficient volume data file.",
-        type="str",
-        default='"sigma_t.vol"',
-    )
-
     weight_filename: t.Optional[str] = documented(
         attr.ib(
             default="weight.vol",
@@ -187,33 +161,9 @@ class ParticleLayer(Atmosphere):
         default='"weight.vol"',
     )
 
-    cache_dir: pathlib.Path = documented(
-        attr.ib(
-            default=pathlib.Path(tempfile.mkdtemp()),
-            converter=pathlib.Path,
-            validator=attr.validators.instance_of(pathlib.Path),
-        ),
-        doc="Path to a cache directory where volume data files will be created.",
-        init_type="path-like",
-        type=":class:`pathlib.Path`",
-        default="temporary directory",
-    )
-
-    def __attrs_post_init__(self):
-        # Prepare cache directory in case we'd need it
-        self.cache_dir.mkdir(parents=True, exist_ok=True)
-
     # --------------------------------------------------------------------------
     #                        Volume data files
     # --------------------------------------------------------------------------
-
-    @property
-    def albedo_file(self) -> pathlib.Path:
-        return self.cache_dir / self.albedo_filename
-
-    @property
-    def sigma_t_file(self) -> pathlib.Path:
-        return self.cache_dir / self.sigma_t_filename
 
     @property
     def weight_file(self) -> pathlib.Path:
@@ -236,6 +186,7 @@ class ParticleLayer(Atmosphere):
             if self.width is not AUTO:
                 warnings.warn(OverriddenValueWarning("Overriding particle layer width"))
             return ctx.override_scene_width
+
         else:
             if self.width is AUTO:
                 spectral_ctx = ctx.spectral_ctx
@@ -585,23 +536,6 @@ class ParticleLayer(Atmosphere):
     #                       Kernel dictionary generation
     # --------------------------------------------------------------------------
 
-    def _gridvolume_to_world_trafo(self, ctx: KernelDictContext) -> t.Any:
-        """
-        Returns the 'to_world' transformation for gridvolume plugins.
-        """
-        length_units = uck.get("length")
-        width = self.kernel_width(ctx).m_as(length_units)
-        top = self.top.m_as(length_units)
-        bottom = self.bottom.m_as(length_units)
-        return map_unit_cube(
-            xmin=-width / 2.0,
-            xmax=width / 2.0,
-            ymin=-width / 2.0,
-            ymax=width / 2.0,
-            zmin=bottom,
-            zmax=top,
-        )
-
     def kernel_phase(self, ctx: KernelDictContext) -> KernelDict:
         particles_phase = self.eval_phase(spectral_ctx=ctx.spectral_ctx)
         return KernelDict(
@@ -609,73 +543,6 @@ class ParticleLayer(Atmosphere):
                 f"phase_{self.id}": {
                     "type": "lutphase",
                     "values": ",".join([str(value) for value in particles_phase.data]),
-                }
-            }
-        )
-
-    def kernel_media(self, ctx: KernelDictContext) -> KernelDict:
-        radprops = self.eval_radprops(spectral_ctx=ctx.spectral_ctx)
-        albedo = to_quantity(radprops.albedo).m_as(uck.get("albedo"))
-        sigma_t = to_quantity(radprops.sigma_t).m_as(uck.get("collision_coefficient"))
-        write_binary_grid3d(
-            filename=str(self.albedo_file), values=albedo[np.newaxis, np.newaxis, ...]
-        )
-        write_binary_grid3d(
-            filename=str(self.sigma_t_file), values=sigma_t[np.newaxis, np.newaxis, ...]
-        )
-        trafo = self._gridvolume_to_world_trafo(ctx=ctx)
-
-        if ctx.ref:
-            phase = {"type": "ref", "id": f"phase_{self.id}"}
-        else:
-            phase = onedict_value(self.kernel_phase(ctx=ctx))
-
-        return KernelDict(
-            {
-                f"medium_{self.id}": {
-                    "type": "heterogeneous",
-                    "phase": phase,
-                    "sigma_t": {
-                        "type": "gridvolume",
-                        "filename": str(self.sigma_t_file),
-                        "to_world": trafo,
-                    },
-                    "albedo": {
-                        "type": "gridvolume",
-                        "filename": str(self.albedo_file),
-                        "to_world": trafo,
-                    },
-                }
-            }
-        )
-
-    def kernel_shapes(self, ctx: KernelDictContext) -> KernelDict:
-        if ctx.ref:
-            medium = {"type": "ref", "id": f"medium_{self.id}"}
-        else:
-            medium = self.kernel_media(ctx)[f"medium_{self.id}"]
-
-        length_units = uck.get("length")
-        width = self.kernel_width(ctx).m_as(length_units)
-        bottom = self.bottom.m_as(length_units)
-        top = self.top.m_as(length_units)
-        offset = self.kernel_offset(ctx).m_as(length_units)
-        trafo = map_cube(
-            xmin=-width / 2.0,
-            xmax=width / 2.0,
-            ymin=-width / 2.0,
-            ymax=width / 2.0,
-            zmin=bottom - offset,
-            zmax=top,
-        )
-
-        return KernelDict(
-            {
-                f"shape_{self.id}": {
-                    "type": "cube",
-                    "to_world": trafo,
-                    "bsdf": {"type": "null"},
-                    "interior": medium,
                 }
             }
         )
